refactor(cli): move getDockerfile to dockerfile.ts, strip build.ts to stub
