+++ packages/cli/src/commands/template/build.ts
@@ -1,12 +1,6 @@
 import * as boxen from 'boxen'
 import * as commander from 'commander'
-import * as fs from 'fs'
-import * as path from 'path'
-import {
-  defaultDockerfileName,
-  fallbackDockerfileName,
-} from 'src/docker/constants'
-import { asBold, asLocalRelative, asPrimary } from '../../utils/format'
+import { asBold, asPrimary } from '../../utils/format'
 
 export const buildCommand = new commander.Command('build')
   .description('Deprecated: use `e2b template create` instead.')
@@ -41,67 +35,3 @@ Migration guide: ${asPrimary('https://e2b.dev/docs/template/migration-v2')}`
     console.log(deprecationWarning)
     process.exit(1)
   })
-
-function loadFile(filePath: string) {
-  if (!fs.existsSync(filePath)) {
-    return undefined
-  }
-
-  return fs.readFileSync(filePath, 'utf-8')
-}
-
-export function getDockerfile(root: string, file?: string) {
-  if (file) {
-    const dockerfilePath = path.join(root, file)
-    const dockerfileContent = loadFile(dockerfilePath)
-    const dockerfileRelativePath = path.relative(root, dockerfilePath)
-
-    if (dockerfileContent === undefined) {
-      throw new Error(
-        `No ${asLocalRelative(
-          dockerfileRelativePath
-        )} found in the root directory.`
-      )
-    }
-
-    return {
-      dockerfilePath,
-      dockerfileContent,
-      dockerfileRelativePath,
-    }
-  }
-
-  let dockerfilePath = path.join(root, defaultDockerfileName)
-  let dockerfileContent = loadFile(dockerfilePath)
-  const defaultDockerfileRelativePath = path.relative(root, dockerfilePath)
-  let dockerfileRelativePath = defaultDockerfileRelativePath
-
-  if (dockerfileContent !== undefined) {
-    return {
-      dockerfilePath,
-      dockerfileContent,
-      dockerfileRelativePath,
-    }
-  }
-
-  dockerfilePath = path.join(root, fallbackDockerfileName)
-  dockerfileContent = loadFile(dockerfilePath)
-  const fallbackDockerfileRelativeName = path.relative(root, dockerfilePath)
-  dockerfileRelativePath = fallbackDockerfileRelativeName
-
-  if (dockerfileContent !== undefined) {
-    return {
-      dockerfilePath,
-      dockerfileContent,
-      dockerfileRelativePath,
-    }
-  }
-
-  throw new Error(
-    `No ${asLocalRelative(defaultDockerfileRelativePath)} or ${asLocalRelative(
-      fallbackDockerfileRelativeName
-    )} found in the root directory (${root}). You can specify a custom Dockerfile with ${asBold(
-      '--dockerfile <file>'
-    )} option.`
-  )
-}
