+++ packages/cli/src/commands/template/dockerfile.ts
@@ -1,0 +1,71 @@
+import * as fs from 'fs'
+import * as path from 'path'
+import {
+  defaultDockerfileName,
+  fallbackDockerfileName,
+} from 'src/docker/constants'
+import { asBold, asLocalRelative } from '../../utils/format'
+
+function loadFile(filePath: string) {
+  if (!fs.existsSync(filePath)) {
+    return undefined
+  }
+
+  return fs.readFileSync(filePath, 'utf-8')
+}
+
+export function getDockerfile(root: string, file?: string) {
+  if (file) {
+    const dockerfilePath = path.join(root, file)
+    const dockerfileContent = loadFile(dockerfilePath)
+    const dockerfileRelativePath = path.relative(root, dockerfilePath)
+
+    if (dockerfileContent === undefined) {
+      throw new Error(
+        `No ${asLocalRelative(
+          dockerfileRelativePath
+        )} found in the root directory.`
+      )
+    }
+
+    return {
+      dockerfilePath,
+      dockerfileContent,
+      dockerfileRelativePath,
+    }
+  }
+
+  let dockerfilePath = path.join(root, defaultDockerfileName)
+  let dockerfileContent = loadFile(dockerfilePath)
+  const defaultDockerfileRelativePath = path.relative(root, dockerfilePath)
+  let dockerfileRelativePath = defaultDockerfileRelativePath
+
+  if (dockerfileContent !== undefined) {
+    return {
+      dockerfilePath,
+      dockerfileContent,
+      dockerfileRelativePath,
+    }
+  }
+
+  dockerfilePath = path.join(root, fallbackDockerfileName)
+  dockerfileContent = loadFile(dockerfilePath)
+  const fallbackDockerfileRelativeName = path.relative(root, dockerfilePath)
+  dockerfileRelativePath = fallbackDockerfileRelativeName
+
+  if (dockerfileContent !== undefined) {
+    return {
+      dockerfilePath,
+      dockerfileContent,
+      dockerfileRelativePath,
+    }
+  }
+
+  throw new Error(
+    `No ${asLocalRelative(defaultDockerfileRelativePath)} or ${asLocalRelative(
+      fallbackDockerfileRelativeName
+    )} found in the root directory (${root}). You can specify a custom Dockerfile with ${asBold(
+      '--dockerfile <file>'
+    )} option.`
+  )
+}
